# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Live Sports Auction Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tournaments/{tournamentId}/teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List teams in a tournament
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teams list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create a team
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /tournaments/{tournamentId}/teams/scout:
    get:
      operationId: getTeamScout
      tags: [teams]
      summary: Scout intelligence — all teams' squads, purses, and unsold players
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scout data with all teams' squads, purse info, and unsold players
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoutData"
        "400":
          description: Invalid tournament ID
  /tournaments/{tournamentId}/teams/{teamId}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get a team by ID
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update a team
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    delete:
      operationId: deleteTeam
      tags: [teams]
      summary: Delete a team
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CATEGORIES ─────────────────────────────────────────────────────────────
...
  /tournaments/{tournamentId}/teams/{teamId}/verify-access:
    post:
      operationId: verifyOwnerAccess
      tags: [teams]
      summary: Verify owner panel access code
      parameters:
        - name: tournamentId
          in: path
          required: true
          schema:
            type: integer
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                required: [valid]
                properties:
                  valid:
                    type: boolean
...
components:
  schemas:
...
    Team:
      type: object
      required: [id, tournamentId, name, shortCode, ownerName, purse, purseUsed, createdAt]
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        name:
          type: string
        shortCode:
          type: string
        ownerName:
          type: string
        ownerMobile:
          type: ["string", "null"]
        ownerEmail:
          type: ["string", "null"]
        ownerPhotoUrl:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        purse:
          type: integer
        purseUsed:
          type: integer
        isBiddingEnabled:
          type: boolean
        accessCode:
          type: ["string", "null"]
        createdAt:
          type: string
    TeamInput:
      type: object
      required: [name, shortCode, ownerName]
      properties:
        name:
          type: string
        shortCode:
          type: string
        ownerName:
          type: string
        ownerMobile:
          type: string
        ownerEmail:
          type: string
        ownerPhotoUrl:
          type: string
        color:
          type: string
        logoUrl:
          type: string
        purse:
          type: integer
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
        shortCode:
          type: string
        ownerName:
          type: string
        ownerMobile:
          type: string
        ownerEmail:
          type: string
        ownerPhotoUrl:
          type: string
        color:
          type: string
        logoUrl:
          type: string
        purse:
          type: integer
        isBiddingEnabled:
          type: boolean
        regenerateCode:
          type: boolean
        reason:
          type: string
          description: Mandatory audit reason for critical team changes (min 10 characters)

    # Scout schemas
    ScoutData:
      type: object
      required: [teams, unsoldPlayers]
      properties:
        teams:
          type: array
          items:
            $ref: "#/components/schemas/ScoutTeam"
        unsoldPlayers:
          type: array
          items:
            $ref: "#/components/schemas/ScoutUnsoldPlayer"
    ScoutTeam:
      type: object
      required: [id, name, purse, purseRemaining, reservePurse, spendablePurse, slotsRequired, playersBought, maximumSquadSize, maxBidCapacity, players]
      properties:
        id:
          type: integer
        name:
          type: string
        shortCode:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        purse:
          type: integer
        purseRemaining:
          type: integer
        reservePurse:
          type: integer
        spendablePurse:
          type: integer
        slotsRequired:
          type: integer
        playersBought:
          type: integer
        maximumSquadSize:
          type: integer
        maxBidCapacity:
          type: integer
        players:
          type: array
          items:
            $ref: "#/components/schemas/ScoutSquadPlayer"
    ScoutSquadPlayer:
      type: object
      required: [id, name, status]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: ["string", "null"]
        status:
          type: string
        soldPrice:
          type: ["integer", "null"]
        isNonPlayingMember:
          type: boolean
    ScoutUnsoldPlayer:
      type: object
      required: [id, name, basePrice]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: ["string", "null"]
        basePrice:
          type: integer
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]

    # Category schemas
...